+++ app/src/main/java/com/gnest/remember/view/fragments/ListItemFragment.java
@@ -497,7 +497,7 @@ public class ListItemFragment extends MvpFragment<IListFragmentView, IListFragme
     }
 
     @Override
-    public void openFromNotification(long id) {
+    public void openFromNotification(int id) {
         presenter.processOpenFromNotification(id);
     }
